LayoutTests/imported/w3c/web-platform-tests/content-security-policy/font-src/font-mismatch-blocked.sub-expected.txt

CONSOLE MESSAGE: Refused to load http://www2.localhost:8800/fonts/Ahem.ttf?font-mismatch-blocked because it does not appear in the font-src directive of the Content Security Policy.

Harness Error (TIMEOUT), message = null

TIMEOUT Test font does not load if it does not match font-src. Test timed out
